鬼剑士技能完成

+++ Assets/Scripts/FixMath/FixIntMath/FixInt.cs
@@ -38,7 +38,7 @@ namespace FixMath
         }
         public FixInt(int value)
         {
-            this.value = value << SHIFT;
+            this.value = (long)value<< SHIFT;
         }
         public FixInt(long value)
         {
